Add files via upload
We will be using sentiment analysis as an example (see SLP3 Chap 4). A machine is asked to predict the sentimental orientation of a piece of texts, such as an Amazon review over a particular product. Supervised learning methods train a classifier, such as SVM and naive Bayes, on a set of reviews that are labeled with sentimental labels (positive and negative), and then predict the sentiments of reviews in the test set.